PIO: implementation is now generic, doesn't contain any target-dependent behavior
Any module can now be connected to any of the PIO ports' pins.
Thus, the VBlank signal is now part of the video module
UART is part of its own module and connects to the PIO on init.
Same goes for the Keyboard.
Zeal's periodic callback can now be delayed.